// frontend/src/services/generated/openapi/model/implementation-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const ImplementationStatus = {
    NotStarted: 'NOT_STARTED',
    Started: 'STARTED',
    WaitingForFeedback: 'WAITING_FOR_FEEDBACK',
    Completed: 'COMPLETED'
} as const;

export type ImplementationStatus = typeof ImplementationStatus[keyof typeof ImplementationStatus];